+++ client/src/views/partials/Footer/Footer.vue
@@ -1,287 +1,81 @@
 <template>
-  <!-- eslint-disable-next-line vue/max-attributes-per-line -->
-  <b-container
-    fluid
-    class="bootstrap bg-white pt-5 pb-5"
-  >
-    <b-container>
-      <b-row class="mb-5">
-        <b-col
-          md="3"
-          cols="6"
-        >
-          <h2>{{ $i18n('footer.call_to_action') }}</h2>
-          <ul class="flex-column">
-            <li class="nav-item">
-              <a
-                :href="$url('infosCompany')"
-                :aria-label="$i18n('footer.infosCompany')"
-              >
-                {{ $i18n('footer.infosCompany') }}
-              </a>
-            </li>
-            <li class="nav-item">
-              <a
-                :href="$url('joininfo')"
-                :aria-label="$i18n('footer.infosPrivatePersons')"
-              >
-                {{ $i18n('footer.infosPrivatePersons') }}
-              </a>
-            </li>
-            <li class="nav-item">
-              <a
-                :href="$url('devdocs_it_tasks')"
-                :aria-label="$i18n('footer.infosDeveloper')"
-              >
-                {{ $i18n('footer.infosDeveloper') }}
-              </a>
-            </li>
-          </ul>
-        </b-col>
-        <b-col
-          md="3"
-          cols="6"
-          class="mb-5 pt-sm-0"
-        >
-          <h2>{{ $i18n(informationen[0].name) }}</h2>
-          <ul class="flex-column">
-            <li
-              v-for="(information, index) in informationen[0].items"
-              :key="index"
-              class="nav-item"
-            >
-              <a
-                :href="$url(information.url)"
-                :aria-label="$i18n(information.title)"
-                v-html="$i18n(information.title)"
-              />
-            </li>
-          </ul>
-        </b-col>
-        <b-col
-          md="6"
-          cols="12"
-          class="pt-xl-5 pt-sm-0"
-        >
-          <b-row>
-            <b-col
-              v-if="isDotAt"
-              xl="6"
-              cols="12"
-            >
-              <p>{{ $i18n('footer.our_partner') }}</p>
-              <a
-                v-b-tooltip="$i18n('footer.hosting_tooltip')"
-                :href="$url('hosting')"
-                class="hide-external"
-                target="_blank"
-                :rel="externalLink"
-              >
-                <img
-                  height="25px"
-                  :alt="$i18n('footer.hosting_tooltip')"
-                  src="/img/footer/manitu_logo.svg"
-                  loading="lazy"
-                >
-              </a>
-              <a
-                :href="$url('denns')"
-                target="_blank"
-                :rel="externalLink"
-              >
-                <img
-                  src="/img/footer/denns_logo_at.png"
-                  loading="lazy"
-                  class="pl-4"
-                  :alt="$i18n('footer.denns_at_alt_text')"
-                >
-              </a>
-              <br>
-              <a
-                :href="$url('bmlfuw')"
-                target="_blank"
-                :rel="externalLink"
-              >
-                <img
-                  src="/img/footer/BMLRT_Logo.png"
-                  loading="lazy"
-                  :alt="$i18n('footer.lebensministerium_alt_text')"
-                  class="pt-4"
-                >
-              </a>
-              <a
-                :href="$url('wiener_tafel')"
-                target="_blank"
-                :rel="externalLink"
-              >
-                <img
-                  src="/img/footer/wiener_tafel_logo_at.png"
-                  loading="lazy"
-                  class="pl-4 pt-4"
-                  :alt="$i18n('footer.wiener_tafel_alt_text')"
-                >
-              </a>
-            </b-col>
-            <b-col
-              v-else
-              sm="6"
-              cols="12"
-              class="mb-5"
-            >
-              <p>{{ $i18n('footer.hosting_info') }}</p>
-              <a
-                v-b-tooltip="$i18n('footer.hosting_tooltip')"
-                :href="$url('hosting')"
-                class="hide-external"
-                target="_blank"
-                :rel="externalLink"
-              >
-                <img
-                  height="49px"
-                  loading="lazy"
-                  class="img-fluid"
-                  :alt="$i18n('footer.hosting_tooltip')"
-                  src="/img/footer/manitu_logo.svg"
-                >
-              </a>
-            </b-col>
-            <b-col
-              sm="6"
-              cols="12"
-              class="pt-xl-0"
-            >
-              <ul>
-                <li>
-                  <p>{{ $i18n('footer.call_to_donate') }}</p>
-                </li>
-                <li>
-                  <b-button
-                    block
-                    variant="primary"
-                    class="w-75"
-                    :href="$url('donations')"
-                  >
-                    {{ $i18n('footer.donate_now') }}
-                  </b-button>
-                </li>
-              </ul>
-            </b-col>
-          </b-row>
-        </b-col>
-      </b-row>
+  <footer class="bg-white">
+    <div class="container py-5">
       <b-row>
         <b-col
-          md="3"
-          cols="6"
-          class="mb-5 mb-xl-0"
-        >
-          <h2>{{ $i18n(informationen[1].name) }}</h2>
-          <ul class="flex-column">
-            <li
-              v-for="(information, index) in informationen[1].items"
-              :key="index"
-              class="nav-item"
-            >
-              <a
-                :href="$url(information.url)"
-                :aria-label="$i18n(information.title)"
-                v-html="$i18n(information.title)"
-              />
-            </li>
-          </ul>
-        </b-col>
-        <b-col
-          md="3"
-          cols="6"
-          class="mb-5 mb-xl-0"
-        >
-          <h2>{{ $i18n(informationen[2].name) }}</h2>
-          <ul class="flex-column">
-            <li
-              v-for="(information, index) in informationen[2].items"
-              :key="index"
-              class="nav-item"
-            >
-              <a
-                :href="$url(information.url)"
-                :aria-label="$i18n(information.title)"
-                v-html="$i18n(information.title)"
-              />
-            </li>
-          </ul>
-        </b-col>
-        <b-col
-          xl="2"
-          md="2"
-          cols="6"
-          class="mb-5 mb-xl-0"
-        >
-          <h2>{{ $i18n(informationen[3].name) }}</h2>
-          <ul class="flex-column">
-            <li
-              v-for="(information, index) in informationen[3].items"
-              :key="index"
-              class="nav-item"
-            >
-              <a
-                :href="$url(information.url)"
-                :aria-label="$i18n(information.title)"
-                v-html="$i18n(information.title)"
-              />
-            </li>
-          </ul>
-        </b-col>
-        <b-col
-          xl="2"
-          md="2"
-          cols="6"
-          class="mb-5 mb-xl-0"
-        >
-          <h2>{{ $i18n(informationen[4].name) }}</h2>
-          <ul class="flex-column">
-            <li
-              v-for="(information, index) in informationen[4].items"
-              :key="index"
-              class="nav-item"
-            >
-              <a
-                :href="$url(information.url)"
-                :aria-label="$i18n(information.title)"
-                v-html="$i18n(information.title)"
-              />
-            </li>
-          </ul>
-        </b-col>
-        <b-col
-          xl="2"
-          md="2"
+          sm="6"
           cols="12"
-          class="mb-5 mb-xl-0"
+          class="mb-5 mb-sm-0"
         >
-          <h2>{{ $i18n(informationen[5].name) }}</h2>
-          <ul class="flex-column">
+          <h2>{{ $i18n(partnerData.title) }}</h2>
+          <div class="d-flex flex-wrap">
+            <a
+              v-for="(item) in partnerData.items"
+              :key="item.infosCompany"
+              v-b-tooltip="$i18n(item.title)"
+              class="partner hide-external"
+              :class="{
+                'alone': partnerData.items.length === 1,
+              }"
+              :href="$url(item.url)"
+              :aria-label="$i18n(item.title)"
+            >
+              <img
+                :alt="$i18n(item.title)"
+                :src="item.img"
+                loading="lazy"
+              >
+              <p
+                v-if="item.info"
+                class="text-muted mb-0"
+                v-html="$i18n(item.info)"
+              />
+            </a>
+          </div>
+        </b-col>
+        <b-col
+          sm="6"
+          cols="12"
+        >
+          <h2>{{ $i18n('footer.donate.call_to') }}</h2>
+          <a
+            class="alert alert-secondary d-flex align-items-center"
+            :href="$url('donations')"
+          >
+            <i class="icon icon--big fas fa-hands-helping mr-3" />
+            {{ $i18n('footer.donate.cta') }}
+          </a>
+        </b-col>
+      </b-row>
+      <div class="line my-5" />
+      <b-row>
+        <b-col
+          v-for="(data) in footerData"
+          :key="data.title"
+          md="3"
+          cols="6"
+        >
+          <h2>{{ $i18n(data.title) }}</h2>
+          <ul>
             <li
-              v-for="(information, index) in informationen[5].items"
-              :key="index"
+              v-for="(item) in data.items"
+              :key="item.infosCompany"
               class="nav-item"
             >
               <a
-                :href="$url(information.url)"
-                :aria-label="$i18n(information.title)"
-                v-html="$i18n(information.title)"
+                :href="$url(item.url)"
+                :aria-label="$i18n(item.title)"
+                v-html="$i18n(item.title)"
               />
             </li>
           </ul>
         </b-col>
       </b-row>
-      <div class="line pt-3 pt-xl-5" />
-      <b-row
-        class="mt-3 justify-content-between"
-      >
+      <div class="line my-5" />
+      <b-row class="justify-content-between">
         <b-col
           md="6"
           cols="12"
-          class="mb-5 mb-md-0"
         >
           <b-row
             class="col"
@@ -312,7 +106,7 @@
             class="col"
           >
             <a
-              v-for="(social, index) in socials"
+              v-for="(social, index) in socialData"
               :key="index"
               v-b-tooltip="social.name"
               :href="$url(social.url + '_'+ (isDotAt ? 'at' : 'de'))"
@@ -336,7 +130,7 @@
             v-if="isBeta || isDev"
             :href="$url('beta_testing_forum')"
           >
-            {{ $i18n('footer.beta_testing_issues') }}
+            {{ $i18n('footer.meta.beta_testing_issues') }}
           </a>
           <a
             v-else
@@ -345,23 +139,27 @@
           />
           <a
             :href="$url('devdocs')"
-            v-html="$i18n('footer.it_devdocs')"
+            v-html="$i18n('footer.meta.it_devdocs')"
           />
           <a
             v-if="version && isBeta"
             class="text-truncate"
             :href="$url('git_revision', version)"
-            v-html="$i18n('footer.version', { version })"
+            v-html="$i18n('footer.meta.version', { version })"
           />
         </b-col>
       </b-row>
-    </b-container>
-  </b-container>
+    </div>
+  </footer>
 </template>
 
 <script>
+// Data
+import SocialData from './Data/SocialData.json'
+import FooterData from './Data/FooterData.json'
+import PartnerData from './Data/PartnerData.json'
+// Mixins
 import RouteCheckMixin from '@/mixins/RouteAndDeviceCheckMixin'
-
 export default {
   mixins: [RouteCheckMixin],
   props: {
@@ -373,112 +171,56 @@ export default {
   data () {
     return {
       externalLink: 'nofollow noreferrer noopener',
-      socials: [
-        {
-          name: 'Twitter',
-          url: 'twitter',
-          icon: 'fab fa-twitter',
-        },
-        {
-          name: 'Facebook',
-          url: 'facebook',
-          icon: 'fab fa-facebook',
-        },
-        {
-          name: 'Instagram',
-          url: 'instagram',
-          icon: 'fab fa-instagram',
-        },
-        {
-          name: 'LinkedIn',
-          url: 'linkedin',
-          icon: 'fab fa-linkedin',
-        },
-        {
-          name: 'YouTube',
-          url: 'youtube',
-          icon: 'fab fa-youtube',
-        },
-        {
-          name: 'TikTok',
-          url: 'tiktok',
-          icon: 'fab fa-tiktok',
-        },
-      ],
-      informationen: [
-        {
-          name: 'menu.entry.regionalgroups',
-          items: [
-            { url: 'communitiesGermany', title: 'menu.entry.Germany' },
-            { url: 'communitiesAustria', title: 'menu.entry.Austria' },
-            { url: 'communitiesSwitzerland', title: 'menu.entry.Swiss' },
-            { url: 'international', title: 'menu.entry.international' },
-          ],
-        },
-        {
-          name: 'menu.entry.aboutUs',
-          items: [
-            { url: 'mission', title: 'menu.entry.mission' },
-            { url: 'wiki_grundsaetze', title: 'menu.entry.fundamentals' },
-            { url: 'blog', title: 'menu.entry.blog' },
-            { url: 'team', title: 'menu.entry.team' },
-            { url: 'partner', title: 'menu.entry.partners' },
-            { url: 'press', title: 'menu.entry.press' },
-          ],
-        },
-        {
-          name: 'menu.entry.background',
-          items: [
-            { url: 'freshdesk', title: 'menu.entry.support', target: '_blank' },
-            { url: 'wiki', title: 'menu.entry.wiki' },
-            { url: 'wiki_guide', title: 'menu.entry.guide', target: '_blank' },
-            { url: 'statistics', title: 'menu.entry.statistics' },
-            { url: 'release_notes', title: 'menu.entry.release-notes' },
-          ],
-        },
-        {
-          name: 'menu.entry.fundraising',
-          items: [
-            { url: 'donations', title: 'menu.entry.donations' },
-            { url: 'circle_of_friends', title: 'menu.entry.friendcircle' },
-            { url: 'selfservice', title: 'menu.entry.selfservice' },
-            { url: 'transparency', title: 'menu.entry.transparency' },
-          ],
-        },
-        {
-          name: 'menu.entry.politics',
-          items: [
-            { url: 'fsstaedte', title: 'menu.entry.fscities' },
-            { url: 'claims', title: 'menu.entry.demands' },
-            { url: 'leeretonne', title: 'menu.entry.pastcampaigns' },
-          ],
-        },
-        {
-          name: 'menu.entry.education',
-          items: [
-            { url: 'academy', title: 'menu.entry.academy' },
-            { url: 'workshops', title: 'menu.entry.talksandworkshops' },
-            { url: 'festival', title: 'menu.entry.fsfestival' },
-          ],
-        },
-      ],
+      socialData: SocialData,
+      footerData: FooterData,
     }
+  },
+  computed: {
+    partnerData () {
+      return this.isDotAt ? PartnerData.at : PartnerData.de
+    },
   },
 }
 </script>
 
 <style lang="scss" scoped>
-.bootstrap .nav-link {
-  padding: unset;
-}
-
 .social_icons {
-  color: var(--fs-color-gray-500);
+  color: var(--fs-color-secondary-900);
   font-size: 1.2rem;
   padding: .25rem;
+  transition: color .2s ease-in-out;
 
   :not(:last-child) {
     margin-right: .5rem;
+  }
+
+  &:hover {
+    color: var(--fs-color-secondary-600);
+  }
+}
+.partner {
+  margin: .5rem;
+
+  img {
+    height: 100%;
+    width: 100%;
+  }
+  &.alone {
+    img {
+      max-width: 156px;
+    }
+  }
+  &:not(.alone) {
+    img {
+      max-width: 75px;
+      @media (max-width: 768px) {
+        max-width:  45px;
+      }
+    }
+  }
+
+  &:first-child {
+    margin-left: 0;
   }
 }
 
@@ -488,24 +230,34 @@ ul {
 }
 
 h2 {
-  font-family: Alfa Slab One,serif;
-  font-style: normal;
-  font-weight: normal;
   font-size: 1.1rem;
-  color: var(--fs-color-gray-600);
 }
 
 .line {
-  border-bottom: 1px solid var(--fs-color-gray-600);
+  border-bottom: 1px solid var(--fs-color-border);
 }
 
-a, p, li {
-  font-size: .8rem;
-  line-height: 1.8em;
-  color: var(--fs-color-gray-900);
-}
+footer {
 
-p {
-  line-height: 1rem;
+  a, p, li {
+    font-size: .8rem;
+    line-height: 1.8em;
+    color: var(--fs-color-dark);
+    text-decoration: none;
+    font-weight: normal;
+  }
+
+  .alert {
+      color: var(--fs-color-secondary-600);
+      font-size: 0.9rem;
+      font-weight: 600;
+      transition: all .2s ease-in-out;
+      line-height: 1.35;
+
+      &:hover {
+        background-color: var(--fs-color-secondary-600);
+        color: var(--fs-color-secondary-100);
+      }
+  }
 }
 </style>